updates for obtaining out dtype
